+++ apps/web/src/app/register-club/_helpers/hooks/useRegisterClubForm.ts
@@ -1,0 +1,28 @@
+import { useForm } from 'react-hook-form';
+
+import { zodResolver } from '@hookform/resolvers/zod';
+import z from 'zod';
+
+export const registerClubSchema = z.object({
+  clubLogoImageUrl: z.string().trim(),
+  clubName: z.string().trim(),
+  clubEnglishName: z.string().trim(),
+  clubDescription: z.string().trim(),
+});
+
+export type RegisterClubFormData = z.infer<typeof registerClubSchema>;
+
+export const useRegisterClubForm = () => {
+  const form = useForm<RegisterClubFormData>({
+    resolver: zodResolver(registerClubSchema),
+    mode: 'onChange',
+    defaultValues: {
+      clubLogoImageUrl: '',
+      clubName: '',
+      clubEnglishName: '',
+      clubDescription: '',
+    },
+  });
+
+  return { form };
+};
